+++ changelog.md
@@ -4,6 +4,8 @@ TGUI 1.8  (TBD)
 - New widget: ContextMenu
 - Added handleWindowEvents function to SFML backend (only for SFML 3)
 - Added onRightClick signal to ListBox
+- Added setSmooth and isSmooth to CanvasSFML
+- Changed the way the view is handled in CanvasSFML
 
 
 TGUI 1.7  (22 December 2024)